Add images for some weather conditions
Added images for "Snow", "Thunderstorm", "Cloudy", "Clear", and "Rain" weather conditions.
