+++ src/App.js
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import "./App.css";
 import logo from "./mlh-prep.png";
 
-import ThemedBackground from "./components/ThemedBackground";
+import ThemedBackground from "./components/theme/ThemedBackground";
 
 function App() {
   const [error, setError] = useState(null);
@@ -64,7 +64,9 @@ function App() {
             )}
           </div>
         </div>
-        <ThemedBackground weatherCondition={results.weather[0].main} />
+        {results && (
+          <ThemedBackground weatherCondition={results?.weather[0].main} />
+        )}
       </>
     );
   }
